# src/main/resources/static/openapi.yaml
openapi: 3.0.3

info:
  title: Policy Overview Dashboard
  description: |
    Read-only and review-management API for insurance policies.
    Contract-first: this specification is the single source of truth for all endpoint shapes.
  version: 1.0.0
  contact:
    email: meet.ravig@gmail.com

servers:
  - url: http://localhost:8081
    description: Local development

tags:
  - name: Policies
    description: Insurance policy retrieval and review management

paths:

  /api/v1/policies:
    get:
      tags: [Policies]
      summary: List policies
      description: |
        Returns a paginated, filterable list of insurance policies.
        All filter parameters are optional. Supports free-text search across
        policyNumber, policyholderName, and underwriter.
      operationId: getPolicies
      parameters:
        - name: status
          in: query
          required: false
          schema: { $ref: '#/components/schemas/PolicyStatus' }
        - name: region
          in: query
          required: false
          schema: { $ref: '#/components/schemas/Region' }
        - name: lineOfBusiness
          in: query
          required: false
          schema: { $ref: '#/components/schemas/LineOfBusiness' }
        - name: effectiveDateFrom
          in: query
          required: false
          description: Filter policies with effective date on or after this date (yyyy-MM-dd).
          schema: { type: string, format: date, example: "2024-01-01" }
        - name: effectiveDateTo
          in: query
          required: false
          description: Filter policies with effective date on or before this date (yyyy-MM-dd).
          schema: { type: string, format: date, example: "2024-12-31" }
        - name: search
          in: query
          required: false
          description: Free-text search across policyNumber, policyholderName, and underwriter (case-insensitive).
          schema: { type: string, example: "Acme" }
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 0, minimum: 0 }
        - name: size
          in: query
          required: false
          schema: { type: integer, default: 10, minimum: 1, maximum: 100 }
        - name: sort
          in: query
          required: false
          description: "Sort field and direction, e.g. sort=premiumAmount,desc"
          schema: { type: string, default: effectiveDate,desc }
      responses:
        '200':
          description: Paginated list returned successfully.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PolicyPage' }
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /api/v1/policies/{id}:
    get:
      tags: [Policies]
      summary: Get policy by ID
      operationId: getPolicyById
      parameters:
        - name: id
          in: path
          required: true
          description: Policy identifier (UUID).
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Policy found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PolicySummaryResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /api/v1/policies/flag:
    patch:
      tags: [Policies]
      summary: Bulk flag policies for review
      description: |
        Sets flaggedForReview = true on all policies whose IDs are in the request array.
        IDs that do not exist are silently ignored. Returns the count actually updated.
      operationId: flagPoliciesForReview
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FlagPoliciesRequest' }
      responses:
        '200':
          description: Flag operation completed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FlagPoliciesResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /api/v1/policies/summary:
    get:
      tags: [Policies]
      summary: Aggregated policy statistics
      description: |
        Returns counts by status, total premium by line of business,
        and the count of active policies expiring within the warning window.
      operationId: getPolicySummary
      responses:
        '200':
          description: Summary statistics returned successfully.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PolicySummaryStats' }
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

components:

  schemas:

    PolicyStatus:
      type: string
      description: Lifecycle status of the policy.
      enum: [ACTIVE, EXPIRED, PENDING, CANCELLED]

    Region:
      type: string
      description: APAC region the policy belongs to.
      enum: [SINGAPORE, HONG_KONG, AUSTRALIA, JAPAN, THAILAND, INDONESIA, MALAYSIA, PHILIPPINES]

    LineOfBusiness:
      type: string
      description: "Line of business. Display names: PROPERTY=Property, CASUALTY=Casualty, ACCIDENT_AND_HEALTH=A&H, MARINE=Marine."
      enum: [PROPERTY, CASUALTY, ACCIDENT_AND_HEALTH, MARINE]

    PolicySummaryResponse:
      type: object
      required: [id, policyNumber, status, premiumAmount, currency, effectiveDate, expiryDate, flaggedForReview, isExpiringSoon]
      properties:
        id:
          type: string
          format: uuid
          example: bd8dcd10-5b1e-44a6-a397-b5820ec83397
        policyNumber:
          type: string
          example: POL-100001
          description: "Unique, format POL-XXXXXX."
        policyholderName:
          type: string
          nullable: true
          example: Wei Chen
        lineOfBusiness:
          type: string
          nullable: true
          example: Property
          description: Human-readable line of business (e.g. A&H). Never the raw enum value.
        status:
          type: string
          example: Active
          description: Title-cased status. Never the raw enum value.
        premiumAmount:
          type: number
          format: decimal
          example: 250000.00
          description: Range 1,000 - 5,000,000.
        currency:
          type: string
          example: SGD
          description: One of USD, SGD, HKD, AUD, JPY, THB.
        effectiveDate:
          type: string
          format: date
          example: "2024-01-01"
        expiryDate:
          type: string
          format: date
          example: "2029-01-01"
        region:
          type: string
          nullable: true
          example: Singapore
          description: Human-readable region name. Never the raw enum value.
        underwriter:
          type: string
          nullable: true
          example: Acme Underwriting Co.
        flaggedForReview:
          type: boolean
          example: false
        isExpiringSoon:
          type: boolean
          example: false
          description: True when expiryDate is within the configured warning window (default 30 days) from today.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    PolicyPage:
      type: object
      required: [content, totalElements, totalPages, size, number]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/PolicySummaryResponse' }
        totalElements: { type: integer, format: int64, example: 220 }
        totalPages: { type: integer, example: 22 }
        size: { type: integer, example: 10 }
        number: { type: integer, example: 0 }
        first: { type: boolean, example: true }
        last: { type: boolean, example: false }

    FlagPoliciesRequest:
      type: object
      required: [policyIds]
      properties:
        policyIds:
          type: array
          minItems: 1
          items: { type: string, format: uuid }
          description: IDs of policies to flag for review. Must contain at least one ID.

    FlagPoliciesResponse:
      type: object
      required: [flaggedCount, policyIds]
      properties:
        flaggedCount:
          type: integer
          example: 3
          description: Number of policies actually updated.
        policyIds:
          type: array
          items: { type: string, format: uuid }

    PolicySummaryStats:
      type: object
      required: [countsByStatus, totalPremiumByLineOfBusiness, expiringSoonCount]
      properties:
        countsByStatus:
          type: object
          additionalProperties: { type: integer, format: int64 }
          example: { Active: 60, Expired: 54, Pending: 53, Cancelled: 53 }
        totalPremiumByLineOfBusiness:
          type: object
          additionalProperties: { type: number, format: decimal }
          example: { Property: 134789786.03, Casualty: 146263987.38, "A&H": 135764529.56, Marine: 139643122.77 }
        expiringSoonCount:
          type: integer
          format: int64
          example: 7

    ErrorResponse:
      type: object
      required: [timestamp, status, message]
      properties:
        timestamp: { type: string, format: date-time, example: "2026-06-07T10:30:00" }
        status: { type: integer, example: 503 }
        message:
          type: string
          example: Policy service is temporarily unavailable. Please try again later.
          description: Human-readable message. Never contains stack traces or internal class names.

  responses:
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { timestamp: "2026-06-07T10:30:00", status: 404, message: "Policy not found with id: bd8dcd10-5b1e-44a6-a397-b5820ec83397" }

    BadRequest:
      description: The request body or parameters failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { timestamp: "2026-06-07T10:30:00", status: 400, message: "policyIds must not be empty" }

    ServiceUnavailable:
      description: Database or downstream service is unavailable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { timestamp: "2026-06-07T10:30:00", status: 503, message: "Policy service is temporarily unavailable. Please try again later." }